Ported cubiomes to C++ (tested on linux using C++17 with Wextra and Werror enabled) and fixed some obvious bugs like missing breaks in switch statements
Note: I have changed as little code as possible and there are lots of optimizations that can be made. Also a lot of
C++ pointer protection that should be added.  I may do that in later commits but I am just hacking this together for the moment to get my seed finder
working.

While not superior to C I think C++ better fits this application as it allows for easy implementation of many algorithms without needing to write
a lot of custom data structures.  I am also a professional modern C++ software architect so I am rather biased to say the least...

Note I plan to port this to CMake in a future commit so it can be better aligned with industry standards.